% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_sbidata_function.R
\name{aswe_sbidata}
\alias{aswe_sbidata}
\title{Function for getting the ASWE data and statistics to calculate SBI values for. SBI function, Ashlee Jollymore 8Feb2021}
\usage{
aswe_sbidata(
  ASWE_sites,
  date_sbi,
  survey_period,
  get_year,
  colnames_data,
  normal_max,
  normal_min,
  force
)
}
\arguments{
\item{ASWE_sites}{ASWE sites that you want to get data and use to calculate SBI data for}

\item{date_sbi}{date that you want to calculate the SBI value for}

\item{survey_period}{Survey period that you want to calculate the SBI for}

\item{get_year}{Water year you want to calculate the SBI for}

\item{colnames_data}{Column names for the final dataframe fo ASWE data that will then be used to calculate SBI data}

\item{normal_max}{Max year of normal period}

\item{normal_min}{Min year of normal period}

\item{force}{Force recalculation of statistics}
}
\description{
Inputs are a list of sites, the date, any incorrect manual sites and the corrected values for these sites.
}
\examples{
aswe_sbidata()
}
\keyword{ASWE}
\keyword{Get}
\keyword{SBI}
\keyword{data}
