% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_sites_function.R
\name{basin_sites}
\alias{basin_sites}
\title{Function for getting the sites within a specific basin}
\usage{
basin_sites(get_basin = "All", exceptions = NULL)
}
\arguments{
\item{get_basin}{Basin that you want to get the snow sites within the basin for. Defaults to "All".}

\item{exceptions}{Stations that you want to exclude. Defaults to NULL}
}
\description{
Inputs are which basin you want to return sites for. Defaults to "All" sites
}
\examples{
basin_sites()
}
\keyword{Basin}
\keyword{basin}
\keyword{by}
\keyword{sites}
