% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{check_write_to_data_dir}
\alias{check_write_to_data_dir}
\title{Function for returning the files within the cached directory}
\usage{
check_write_to_data_dir(dir, ask)
}
\arguments{
\item{dir}{data firectory}

\item{ask}{Whether to ask the user whether the program can create a directory if it doesn't exist.}
}
\description{
Function for returning the files within the cached directory
}
\examples{
check_write_to_data_dir()
}
\keyword{data}
\keyword{dir}
\keyword{files}
