% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{data_dir}
\alias{data_dir}
\title{Function for returning the correct data cache directory}
\usage{
data_dir()
}
\description{
Function for returning the correct data cache directory
}
\examples{
data_dir()
}
\keyword{data}
\keyword{dir}
