% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_function.R
\name{elevation_data}
\alias{elevation_data}
\title{Function for returning elevation by station}
\usage{
elevation_data()
}
\description{
Function for returning elevation by station
}
\examples{
\dontrun{}
}
\keyword{internal}
