% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sbi_year_function.R
\name{get_SBI_year}
\alias{get_SBI_year}
\title{Function that gets data to calculate SBI and then calculates the SBI using the manual and ASWE data retrieved.}
\usage{
get_SBI_year(date_sbi, sites, incorrect_sites, incorrect_data)
}
\arguments{
\item{date_sbi}{Date that you want to calculate the SBI value for}

\item{sites}{Sites that you want to get data and use to calculate SBI data for}

\item{incorrect_sites}{Manual site IDs that are incorrect within current survey date}

\item{incorrect_data}{Data for the manual sites that are incorrect - this is the corrected data that the suer wants to replace with}
}
\description{
This function collects the statistics data for individual sites using the get_SBI_data() function, and then calculates the SBI by basin using SBI_sites_function()
}
\examples{
\dontrun{}
}
\keyword{internal}
