% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_sbi_data_function.R
\name{manual_sbi_data}
\alias{manual_sbi_data}
\title{Function for assembling manual data for SBI calculation}
\usage{
manual_sbi_data(
  survey_period,
  manual_sites,
  get_year,
  normals_manual,
  colnames_data_manual,
  normal_max,
  normal_min,
  incorrect_sites = NA,
  incorrect_data = NA,
  force = FALSE
)
}
\arguments{
\item{survey_period}{Survey period you are calculating SBI for}

\item{manual_sites}{Manual sites that you want to get data and use to calculate SBI data for}

\item{get_year}{Year that you are calculating the SBI for}

\item{normals_manual}{previous normals calculated for manual sites. This is data within the /data folder}

\item{colnames_data_manual}{Column names of the manual data that you will eventually export}

\item{normal_max}{Date (year) of max year for normal period}

\item{normal_min}{Date (year) of min year for normal period}

\item{force}{whether to force the recalculation of manual normals. Defaults to FALSE (no)}
}
\description{
Function calculates statistics for the manual sites that you specify for the specific year and survey period
}
\examples{
\dontrun{}
}
\keyword{internal}
