% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SBI_data_function.R
\docType{data}
\name{normals_ASWE}
\alias{normals_ASWE}
\title{Function for aggregating all of the data needed to calculate the SBI for a basin.}
\format{
An object of class \code{data.frame} with 71 rows and 20 columns.
}
\usage{
normals_ASWE
}
\arguments{
\item{sites}{Sites that you want to get data and use to calculate SBI data for}

\item{date_sbi}{date that you want to calculate the SBI value for}

\item{incorrect_sites}{Manual site IDs that are incorrect within current survey date}

\item{incorrect_data}{Data for the manual sites that are incorrect - this is the corrected data that the suer wants to replace with}

\item{normals_ASWE}{previously calculated normals. defaults to bcsnowstats::ASWE_normals_1981t2010}

\item{normals_manual}{Previously calculated normals. Defaults to bcsnowstats::manual_normals_1981t2010}
}
\description{
Inputs are a list of sites, the date, any incorrect manual sites and the corrected values for these sites.
}
\examples{
\dontrun{}
}
\keyword{internal}
