% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_function.R
\name{round_df}
\alias{round_df}
\title{function for rounding all numerics within a dataframe to a certain significant figure}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{data}

\item{digits}{number of digits that you want to round to}
}
\description{
function for rounding all numerics within a dataframe to a certain significant figure
}
\examples{
round_df()
}
\keyword{dataframe}
\keyword{entries}
\keyword{round}
