% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBI_function.R
\name{SBI_function}
\alias{SBI_function}
\title{Function for calculating the SBI once the data is collected for individual sites (output from get_SBI_data() function)}
\usage{
SBI_function(data, date_sbi, basins)
}
\arguments{
\item{data}{dataframe of sites that you are looking to calculate the SBI for}

\item{date_sbi}{Date you are calculating the SBI for}

\item{basins}{Basins that you want to calculate SBI for}
}
\description{
Function for calculating the SBI once the data is collected for individual sites (output from get_SBI_data() function)
}
\examples{
SBI_function()
}
\keyword{SBI}
\keyword{calculation}
