% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBI_sites_function.R
\name{SBI_sites_function}
\alias{SBI_sites_function}
\title{Script for calculating the SBI for one basin. This is the function called within a lapply loop to go over multiple basins.
Meant to be called after get_SBI_data() within "SWE_SBI_getdata_function.R".
This function will call the SBI_function() below to calculate the SBI for a particular basin.}
\usage{
SBI_sites_function(basins_initial, date_sbi, SBI_data)
}
\arguments{
\item{basins_initial}{initial basins you want to calculate SBI for}

\item{date_sbi}{Date of SBI}

\item{SBI_data}{Site data that you are using to calculate SBI for}

\item{data}{dataframe of sites that you are looking to calculate the SBI for}
}
\description{
Script for calculating the SBI for one basin. This is the function called within a lapply loop to go over multiple basins.
Meant to be called after get_SBI_data() within "SWE_SBI_getdata_function.R".
This function will call the SBI_function() below to calculate the SBI for a particular basin.
}
\examples{
SBI_sites_function()
}
\keyword{SBI}
\keyword{calculation}
