% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{show_cached_files}
\alias{show_cached_files}
\title{Function for returning the files within the cached directory}
\usage{
show_cached_files()
}
\description{
Function for returning the files within the cached directory
}
\examples{
show_cached_files()
}
\keyword{data}
\keyword{dir}
\keyword{files}
