% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_basinname_function.R
\name{site_basinname}
\alias{site_basinname}
\title{Function for getting the snow basin name for a specific site (or all, or all active sites)}
\usage{
site_basinname(id = "All")
}
\arguments{
\item{id}{Station ID that you want to get the basin name for. Defaults to "All", but can also be a string of sites, one site, or "active" sites.}
}
\description{
Inputs are a list of sites,  "All" (that returns the basin name for all of the sites), or "active", which returns the basin name for all the current active sites
}
\examples{
site_basinname()
}
\keyword{Get}
\keyword{ID}
\keyword{basin}
\keyword{by}
\keyword{name}
