% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowbasinaverage_function.R
\name{snow_basin_average}
\alias{snow_basin_average}
\title{Function for returning the SBI for a specific subset of sites that you manually determine}
\usage{
snow_basin_average(
  sites,
  date_sbi,
  incorrect_sites = NA,
  incorrect_data = NA,
  force = FALSE,
  ask = FALSE,
  ...
)
}
\arguments{
\item{sites}{String of site IDs that you want to calculate SBI values for}

\item{incorrect_sites}{Site ID's for manual sites with incorrect data. Defaults to NA.}

\item{incorrect_data}{Corrected data for manual sites that have incorrect data. Defaults to NA.}

\item{force}{Whether all of the data should be recalculated rather than retrieved from the statistics data cache. Defaults to FALSE or no.}

\item{ask}{Whether to ask the user to create a data cache directory. Defaults to FALSE or no.}

\item{date_SBI}{date of the SBI you want to retrieve ("YYYY-mm-dd")}
}
\description{
Function for returning the SBI for a specific subset of sites that you manually determine
}
\examples{
snow_basin_average()
}
\keyword{Get}
\keyword{SBI}
\keyword{data}
\keyword{manual}
