% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwind_SBI}
\alias{unwind_SBI}
\title{Functions for unwinding lists created by SBI_sites_function()}
\usage{
unwind_SBI(list)
}
\description{
Functions for unwinding lists created by SBI_sites_function()
}
\examples{
\dontrun{}
}
\keyword{internal}
