% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbi_bybasin_function.R
\name{sbi_bybasin_function}
\alias{sbi_bybasin_function}
\title{SBI Function. For calculating SBI by basin. Function for calculating and saving the SBI by year.
Select a year, a survey period (can be all), and this function calculates the SBI and the sites statistics and saves both as two csv files.}
\usage{
sbi_bybasin_function(
  date_sbi,
  all_basins = "Yes",
  exceptions = NA,
  incorrect_sites = NA,
  incorrect_data = NA,
  save_csv = c("No"),
  path
)
}
\arguments{
\item{date_sbi}{Date that you want to calculate SBI for. Must be in date format (i.e., as.Date("YYYY-MM-DD"))}

\item{all_basins}{What basin you want to calculate the SBI for. Defaults to "Yes", which will return the SBI for all of the current snow basins.}

\item{exceptions}{Any sites that you want to exclude from the analysis. Defaults to NA}

\item{incorrect_sites}{Manual site IDs that are showing incorrect or suspect SWE values. Defaults to NA.}

\item{incorrect_data}{a string of numeric values that are the correct values for the incorrect sites identified in incorrect_sites. Defaults to NA.}

\item{save_csv}{Whether the user wants to save a csv file of the final SBI values as well as the statistics from the sites used to calculate the SBI values. Can be 'Yes' or "No", but defaults to "No".}

\item{path}{Path that you want to save the SBI result to}

\item{force}{Whether you want to force the fresh download of the statistics for the sites used to calculate SBI values. Defaults to FALSE}

\item{ask}{Whether to ask the user if they want to create a cached data directory. Defaults to FALSE}

\item{use_sbi_cache}{Whether to try and retrieve SBI values from the cache. If the data exists, then the SBI values will NOT be recalculated. Defaults to FALSE}
}
\description{
This function gives you the SBI for one, multiple, or all basins by water year for a specific day or survey period.
}
\examples{
sbi_bybasin_function()
}
\keyword{SBI}
